+++ Command/EditPetsCommand.cs
@@ -3,7 +3,7 @@ using PetPals_BackEnd_Group_9.Models;
 
 namespace PetPals_BackEnd_Group_9.Command
 {
-    public record EditPetsCommand(int PetId, string Name, string Breed, int Age, string Description, decimal Price) : IRequest<EditPetsResult>;
+    public record EditPetsCommand(int PetId, string Name, string Breed, decimal Age, string Description, decimal Price) : IRequest<EditPetsResult>;
 
 
 }
