+++ Models/GetOwnerPetsResponse.cs
@@ -6,7 +6,7 @@
         public string Name { get; set; } = string.Empty;
         public string Slug { get; set; } = string.Empty;
         public string Breed { get; set; } = String.Empty;
-        public int Age { get; set; }
+        public decimal Age { get; set; }
         public string Gender { get; set; } = string.Empty;
         public string Description { get; set; } = String.Empty;
         public string Status { get; set; } = String.Empty;
