+++ Models/PetResponse.cs
@@ -5,7 +5,7 @@
         public int PetId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Breed { get; set; } = string.Empty;
-        public int Age { get; set; }
+        public decimal Age { get; set; }
         public string Gender { get; set; } = string.Empty;
         public string Species { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
